requirements.txt
django
bootstrap4
django-bootstrap4
psycopg2
djangorestframework
djangorestframework-simplejwt
djangorestframework-jwt
